Parse low, hight and duration as floats

+++ DeWave/cmddataprep.py
@@ -8,11 +8,11 @@ parser.add_argument("--dir", type=str, help="input folder name",
 default="data")
 parser.add_argument("--num", type=int, help="The number of clips for\
 each speaker", default=128)
-parser.add_argument("--duration", type=int, help="The duration of each \
+parser.add_argument("--duration", type=float, help="The duration of each \
 clip", default=5)
-parser.add_argument("--low", type=int, help="starting time of the audio from \
+parser.add_argument("--low", type=float, help="starting time of the audio from \
 which the clip is sampled", default=0)
-parser.add_argument("--high", type=int, help="ending time of the audio from \
+parser.add_argument("--high", type=float, help="ending time of the audio from \
 which the clip is sampled", default=600)
 parser.add_argument("--out", type=str, help="the output directory",
 default="data/train")
